I15_1-407 (9.3pre) Rename queues.spooler plugin to queuesspooler
Since org.eclipse.scanning.event.queues is defined in another plugin,
org.eclipse.scanning.event.queues.spooler.ServicesHolder cannot be found
in Jython. Moving it into it's own packaged fixes this problem.
