+++ org.eclipse.scanning.event.queuesspooler/.classpath
