+++ org.eclipse.scanning.event.queuesspooler/.gitignore
