+++ org.eclipse.scanning.event.queuesspooler/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.eclipse.scanning.event.queues.spooler</name>
+	<name>org.eclipse.scanning.event.queuesspooler</name>
 	<comment></comment>
 	<projects>
 	</projects>
