+++ org.eclipse.scanning.event.queuesspooler/.settings/org.eclipse.jdt.core.prefs
