+++ org.eclipse.scanning.event.queuesspooler/META-INF/MANIFEST.MF
@@ -1,7 +1,7 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Queue Spooler
-Bundle-SymbolicName: org.eclipse.scanning.event.queues.spooler;singleton:=true
+Bundle-SymbolicName: org.eclipse.scanning.event.queuesspooler;singleton:=true
 Bundle-Version: 1.0.0.qualifier
 Bundle-Vendor: Diamond Light Source
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
@@ -10,7 +10,7 @@ Import-Package: org.osgi.framework;version="1.7.0",
  org.osgi.service.component;version="1.2.0",
  org.slf4j;version="[1.7.2,2.0.0)"
 Service-Component: OSGI-INF/servicesHolder.xml
-Export-Package: org.eclipse.scanning.event.queues.spooler,
- org.eclipse.scanning.event.queues.spooler.beanassemblers,
- org.eclipse.scanning.event.queues.spooler.modelassemblers
+Export-Package: org.eclipse.scanning.event.queuesspooler,
+ org.eclipse.scanning.event.queuesspooler.beanassemblers,
+ org.eclipse.scanning.event.queuesspooler.modelassemblers
 Bundle-ActivationPolicy: lazy
