+++ org.eclipse.scanning.event.queuesspooler/OSGI-INF/servicesHolder.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="start" deactivate="stop" immediate="true" name="Queue Spooler Service and subsystems microservice Holder">
-   <implementation class="org.eclipse.scanning.event.queues.spooler.ServicesHolder"/>
+   <implementation class="org.eclipse.scanning.event.queuesspooler.ServicesHolder"/>
    <reference bind="setRunnableDeviceService" cardinality="0..1" interface="org.eclipse.scanning.api.device.IRunnableDeviceService" name="IRunnableDeviceService" policy="dynamic" unbind="unsetRunnableDeviceService"/>
    <reference bind="setScannableDeviceService" cardinality="0..1" interface="org.eclipse.scanning.api.device.IScannableDeviceService" name="IScannableDeviceService" policy="dynamic" unbind="unsetScannableDeviceService"/>
 </scr:component>
