+++ org.eclipse.scanning.event.queuesspooler/build.properties
