+++ org.eclipse.scanning.event.queuesspooler/src/org/eclipse/scanning/event/queuesspooler/QueueBeanFactory.java
@@ -1,4 +1,4 @@
-package org.eclipse.scanning.event.queues.spooler;
+package org.eclipse.scanning.event.queuesspooler;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -17,12 +17,12 @@ import org.eclipse.scanning.api.event.queues.models.ExperimentConfiguration;
 import org.eclipse.scanning.api.event.queues.models.QueueModelException;
 import org.eclipse.scanning.api.event.queues.models.arguments.IQueueValue;
 import org.eclipse.scanning.api.event.queues.models.arguments.QueueValue;
-import org.eclipse.scanning.event.queues.spooler.beanassemblers.IBeanAssembler;
-import org.eclipse.scanning.event.queues.spooler.beanassemblers.MonitorAtomAssembler;
-import org.eclipse.scanning.event.queues.spooler.beanassemblers.PositionerAtomAssembler;
-import org.eclipse.scanning.event.queues.spooler.beanassemblers.ScanAtomAssembler;
-import org.eclipse.scanning.event.queues.spooler.beanassemblers.SubTaskAtomAssembler;
-import org.eclipse.scanning.event.queues.spooler.beanassemblers.TaskBeanAssembler;
+import org.eclipse.scanning.event.queuesspooler.beanassemblers.IBeanAssembler;
+import org.eclipse.scanning.event.queuesspooler.beanassemblers.MonitorAtomAssembler;
+import org.eclipse.scanning.event.queuesspooler.beanassemblers.PositionerAtomAssembler;
+import org.eclipse.scanning.event.queuesspooler.beanassemblers.ScanAtomAssembler;
+import org.eclipse.scanning.event.queuesspooler.beanassemblers.SubTaskAtomAssembler;
+import org.eclipse.scanning.event.queuesspooler.beanassemblers.TaskBeanAssembler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
