+++ org.eclipse.scanning.event.queuesspooler/src/org/eclipse/scanning/event/queuesspooler/beanassemblers/ScanAtomAssembler.java
@@ -1,4 +1,4 @@
-package org.eclipse.scanning.event.queues.spooler.beanassemblers;
+package org.eclipse.scanning.event.queuesspooler.beanassemblers;
 
 import java.lang.reflect.Method;
 import java.util.ArrayList;
@@ -26,11 +26,10 @@ import org.eclipse.scanning.api.event.scan.ScanRequest;
 import org.eclipse.scanning.api.points.models.CompoundModel;
 import org.eclipse.scanning.api.points.models.IScanPathModel;
 import org.eclipse.scanning.api.scan.ScanningException;
-import org.eclipse.scanning.event.queues.spooler.ServicesHolder;
-//import org.eclipse.scanning.event.queues.ServicesHolder;
-import org.eclipse.scanning.event.queues.spooler.modelassemblers.ArrayModelAssembler;
-import org.eclipse.scanning.event.queues.spooler.modelassemblers.IScanObjectModelAssembler;
-import org.eclipse.scanning.event.queues.spooler.modelassemblers.StepModelAssembler;
+import org.eclipse.scanning.event.queuesspooler.ServicesHolder;
+import org.eclipse.scanning.event.queuesspooler.modelassemblers.ArrayModelAssembler;
+import org.eclipse.scanning.event.queuesspooler.modelassemblers.IScanObjectModelAssembler;
+import org.eclipse.scanning.event.queuesspooler.modelassemblers.StepModelAssembler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
